+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/structures/StructuresNodeDialog.java
@@ -1,0 +1,10 @@
+package nl.esciencecenter.e3dchem.gpcrdb.structures;
+
+import nl.esciencecenter.e3dchem.gpcrdb.GpcrdbNodeDialog;
+
+public class StructuresNodeDialog extends GpcrdbNodeDialog {
+	protected StructuresNodeDialog() {
+		super();
+		addAdvancedSettings();
+	}
+}